3.1.12 - usunięto plik z klasą, która nie musi być publiczna (Node)
