+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -43,3 +43,38 @@ public class IntLinkedList {
 	}
 
 }
+
+class Node {
+
+	private int value;
+	private Node prev;
+	private Node next;
+
+	Node(int i) {
+		setValue(i);
+	}
+
+	public int getValue() {
+		return value;
+	}
+
+	public void setValue(int value) {
+		this.value = value;
+	}
+
+	public Node getPrev() {
+		return prev;
+	}
+
+	public void setPrev(Node prev) {
+		this.prev = prev;
+	}
+
+	public Node getNext() {
+		return next;
+	}
+
+	public void setNext(Node next) {
+		this.next = next;
+	}
+}
